requirements.txt
playwright
beautifulsoup4
fastapi
uvicorn
pydantic
python-dotenv
pyyaml
httpx
